Refactor network rule compile helpers
